fix bug and create python file of ProjectClass
